Clean up main.py by creating the tools.py file
